Add my script, now :P (and minor fixes) You can now use WorldEdit to generate a map (But, you can only generate a map full of grass) But, the map can't be used with Server and Client (boost::serialization)
Update TODO
Add my script, it's weirder than the Krofna one :D
Fix test_db_full.sql
(And some uint32 -> uint64)

+++ server/Creature.hpp
@@ -28,7 +28,7 @@ class CreatureAI;
 class WOD_DLL_DECL Creature : public Unit
 {
 public:
-    Creature(uint32 ObjID, MapPtr pMap, uint16 x, uint16 y, CreatureTemplate* pTemplate);
+    Creature(uint64 ObjID, MapPtr pMap, uint16 x, uint16 y, CreatureTemplate* pTemplate);
     ~Creature();
     void InitializeAI();
